+++ LethalMon/Data.cs
@@ -9,19 +9,19 @@ public static class Data
 {
     public static readonly Random Random = new Random();
     
-    public static readonly double[][] CaptureProbabilities =
+    public static readonly float[][] CaptureProbabilities =
     {
         // Normal ball
-        new[] { 0.95, 0.9, 0.8, 0.65, 0.5, 0.3, 0.1, 0.0, 0.0, 0.0 },
+        new[] { 0.95f, 0.9f, 0.8f, 0.65f, 0.5f, 0.3f, 0.1f, 0.0f, 0.0f, 0.0f },
         
         // Great ball
-        new[] { 1.0, 0.97, 0.95, 0.85, 0.7, 0.5, 0.3, 0.2, 0.15, 0.1 },
+        new[] { 1.0f, 0.97f, 0.95f, 0.85f, 0.7f, 0.5f, 0.3f, 0.2f, 0.15f, 0.1f },
         
         // Ultra ball
-        new[] { 1.0, 1.0, 1.0, 0.97, 0.95, 0.9, 0.8, 0.7, 0.6, 0.5 },
+        new[] { 1.0f, 1.0f, 1.0f, 0.97f, 0.95f, 0.9f, 0.8f, 0.7f, 0.6f, 0.5f },
         
         // Master ball
-        new[] { 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0 }
+        new[] { 1.0f, 1.0f, 1.0f, 1.0f, 1.0f, 1.0f, 1.0f, 1.0f, 1.0f, 1.0f }
     };
 
     public static readonly Dictionary<string, CatchableEnemy.CatchableEnemy> CatchableMonsters = new()
